backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=24.0.0",
    "influxdb>=5.3.2",
    "minimalmodbus>=2.1.1",
    "pyjwt>=2.10.1",
    "pymodbus>=3.11.4",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "pyserial>=3.5",
]
